Move fasm2bels to a seperate module.
