+++ xc7/fasm2bels/__init__.py
@@ -1,0 +1,4 @@
+from .fasm2bels import main
+
+if __name__ == "__main__":
+    main()
