+++ xc7/fasm2bels/prjxray_bram_models.py
@@ -1,5 +1,5 @@
 import fasm
-from verilog_modeling import Bel, Site
+from .verilog_modeling import Bel, Site
 
 def get_init(features, target_feature, invert, width):
     init = 0
