+++ xc7/fasm2bels/verilog_modeling.py
@@ -1,6 +1,48 @@
 import functools
-from prjxray_make_routes import make_routes, ONE_NET, ZERO_NET, prune_antennas
-from lib.connection_database import get_wire_pkey
+from .make_routes import make_routes, ONE_NET, ZERO_NET, prune_antennas
+
+def get_wire_pkey(conn, tile_name, wire):
+    c = conn.cursor()
+    c.execute("""
+WITH selected_tile(tile_pkey, tile_type_pkey) AS (
+  SELECT
+    pkey,
+    tile_type_pkey
+  FROM
+    tile
+  WHERE
+    name = ?
+)
+SELECT
+  wire.pkey
+FROM
+  wire
+WHERE
+  wire.tile_pkey = (
+    SELECT
+      selected_tile.tile_pkey
+    FROM
+      selected_tile
+  )
+  AND wire.wire_in_tile_pkey = (
+    SELECT
+      wire_in_tile.pkey
+    FROM
+      wire_in_tile
+    WHERE
+      wire_in_tile.name = ?
+      AND wire_in_tile.tile_type_pkey = (
+        SELECT
+          tile_type_pkey
+        FROM
+          selected_tile
+      )
+  );
+""", (tile_name, wire))
+
+    results = c.fetchone()
+    assert results is not None, (tile_name, wire)
+    return results[0]
 
 
 class Bel(object):
